feat: make mypy precommit platform-agnostic
The difficulty is that the location of the python executable of the
virtual environment (alias or otherwise) is located in a different
place only on Windows. This means the python_executable argument
of mypy must be different depending on platform. pre-commit doesn't
support dynamic args to the hook, so changed it to be a python script
that runs mypy instead. This script gets the correct location depending
on platform and uses it to call mypy. When running mypy locally, this
argument is unnecessary as it will automatically use the python executable
of the current virtual environment (pre-commit has a separate one,
necessitatin the workaround)

Pros:
- platform-agnostic mypy pre-commit hook

Cons:
- no colors in mypy hook output (still there for local mypy usage)